requirements.txt
Flask
dotenv
flask_sqlalchemy
pymysql
cryptography
flask_jwt_extended
flask_cors
python-json-logger